refactor(api): extrai validators search + embedding/rerank de validation.ts
Decomposição god-file (Tier 1 / #3501), 5ª fatia. Move os validators de web-search
e de embedding/rerank/clarifai para módulos co-localizados, byte-idênticos ao
inline original (funções top-level, sem captura de estado do dispatcher).

- providers/validation/searchProviders.ts   — validateGenericProvider, validateSearchProvider,
  SEARCH_VALIDATOR_CONFIGS (serper, tavily, jina-reader, …)
- providers/validation/embeddingProviders.ts — clarifai, embedding-api, rerank-api

validation.ts: 1771 → 1435 (−336). Host importa validateSearchProvider +
SEARCH_VALIDATOR_CONFIGS e os 3 validators de embedding p/ o dispatcher. Removidos:
imports headers órfãos (withCustomUserAgent, buildClarifaiHeaders) + um comentário
órfão sobre muse-spark-web (código já movido na fatia 2).

Validação: typecheck:core, eslint (0/0), check:cycles, check:file-size (módulos
<800), 147 testes de validação existentes + 3 de caracterização — todos verdes.
